feature(news) Install aldryn-newsblog and create templates
This plugin will be used for our news. I have to modify the templates to be usable with fun-mooc css
The templates have the original Fun Mooc html structure.
I added the news blog social part in the blue square of detail view.
It consist of categories list and related articles plugins.