config(settings): set location for static and templates
this commit defines a /static and a /media folders at
the root of the project that are not committed and
that are used for serving files in a development
environment. The /fun_cms/static folder is where
we store statics for the project. /fun_cms/templates
is where the templates for the project will be located.