Build our CSS through SASS so we can structure it properly
Using SASS allows us to easily follow BEM as we write our CSS classes,
lets us modularize our CSS and keep it close to the related markup, and
integrate bootstrap properly by only using the parts we need.

We needed to upgrade Bootstrap to V4 for this as V3 was written in LESS
and we would have had to change things soon after we built them.