Render React components on page load based on a class convention
Before we can work on our components, we need to specify a way React
can recognize the parts of the page it must take control of and know
which components to render there. That's what we do with this detector
in index.tsx. We made a dummy search component to showcase the way it
will work.