Fix the SASS build watcher
SASS needs us to declare any and all folders it should watch, not only
the entry point which imports our other files. This means we need to
watch the entire app folder. As a result we need to prefix our filenames
with _ so sass knows they're partials and don't need to have their own
CSS file generated.