feat(search): implement the search template
The Search page is a specific page with its own layout. It thus needs its
own template in order to use the full width of the page, for instance.
This brings us to drop some assertions from base.html, such as the fact that
any and all content will be held in a regular `.container` element