feat(search): Add API response fixtures & document their types
For now we're only working on the interface, and ignoring any interface
with the server. Commit an API response example so we can easily iterate
on the UI with real data.
The types, on the other hand, are here to stay and will have to evolve as
the underlying data on the backend changes shapes.