conf(build) put generated JS/CSS in source static files
We need to put these files in the sources (as opposed to the /data
folder) so they can be loaded by runserver, and we actually get them
during development.
We'll depend on the collecstatic command to put them in the /data folder
so nginx can access them when it is deployed.