feat(http) add a practical XHR wrapper
While we do not need the XHR wrapper right now, it makes a good showcase
for our new support of unit tests, and we will need it later anyway.
We don't want to use XHR directly as its unwieldiness would force us
to write a lot of boilerplate all over. It can be easily extended when
we need it.