🔧(build) add cache invalidation mechanism for static files
Each time static files are updated but keep the same name,
browsers will get an outdated version from the webserver or CDN
cache which is difficult to invalidate. By including a hash in
the filename, that is calculated from the file content, we make
sure browsers always get the updated version.