feat(search) populate redux state from our fixtures
Now that we removed the direct inclusion from our data service, we
can load the fixtures more properly at bootstrap time. This is convenient
because it will provide us a way to load django-provided data that is
already in the page as well, when/if we decide to implement that.