feature(lint) add pylint CI job
Add pylint dependency and configuration file.

Update the CircleCI workflow so that lint and test jobs all depend only from the
build. A non-linear workflow may potentially speed up CI execution (parallel
tasks) and also raise test failures when the lint job also fails (this wasn't
the case in the previous implementation).

The lint job has been splitted in two separated jobs on purpose: lint-flake8 and
lint-pylint. By doing so, we can choose to accept a PR even if the lint-pylint
job failed (we need to configure the develop branch as such).