fix(docker) force container/local user mapping
It is often considered as a good practice to define a USER in your
Dockerfile to prevent privilege escalation in a Docker context. This can
be an issue if you mount host volumes that must have write permission
for the container's user.

In this case, it is a good idea to force the mapping with local users
thanks to the --user option of docker(-compose) run/exec. In our
particular case, we use this option in the Makefile only for the node
service commands. By doing so, we have access to current user ids ($(id
-u):$(id -g)) and this trick will hopefully work on most systems.