😎(elasticsearch) refactor regenerate_index command into a module
We needed to rework the regenerate_index command so it actually performs
indexation. We needed to make it operate from classes that provide all
the necessary information to generate one index for one type of document,
with the relevant mapping and ES-specific keys.
As we worked on this, we realized it was smarter (and simpler) to put the
actual logic handling the indexing in a separate module and have the command
simply use this module.