😎(search)write a saga to manage organization list requests
In our filters list, organizations will be taken from the facets
of ElasticSearch query responses. However, we'll still need to get the
human names from the server through a list request on an organization
endpoint.
This saga provides the facilities to handle that use case: the relevant
component will just need to dispatch an action with the necessary keys
and the consumer will have the human names ready as soon as they're loaded
from the server.