👷‍♀️(search) split SearchFilter from SearchFilterGroup and add tests
We're in the process of writing smoke tests for all our components. As we
started to do that for SearchFilterGroup we noticed it would be much
easier if the functionality for a single filter was split out from the list.
This commit addresses that separation of concerns and includes the newly
written tests.