😎(search) implement a custom "Array of {FormField}" form field
Implement a custom Django form field type to validate and clean
arrays of values, validating both the array itself and each of
the items it contains, through the appropriate Django form field
type (such as CharField or IntegerField).
This will be useful for some API request query params that can
pass arrays of UUIDs or slugs to filter searches or return an
exact matching list.