✨(search) rework the org indexer to fetch data from the API
When we wrote the OrganizationIndexer, we were counting on the
OrganizationPage model to do the job of importing organizations from
openedx for us. Unfortunately that part is not ready yet; in order to
move on, we're doing it directly in the indexer for not, just like
we do for subjects & courses.