👷(circle) move to a richie's container based CI
We now use CircleCI to build Richie's docker images that will be used in
various environments:

* production:
    - fundocker/richie:latest
    - fundocker/richie:alpine
* test/dev:
    - fundocker/richie:${VERSION}-dev
    - fundocker/richie:${VERSION}-alpine-dev

Once built, we use those images to run python linters and the test suite
on it.

If our images have been built successfully, and all tests are green, we
upload those images to DockerHub for each new git tag pushed to the
github repository.

For now, front-end jobs are treated separately using a node-based image.