♻️(front) factorize "resourceById" reducer in /genericReducers
This chunk of reducer logic that handles the byId slice of the
state for a given resource was common to all 3 of of resource
types. It's simple enough that it did not seem harmful but from
 a correctness standpoint it should not be repeated 3 times.
Factorize it, it will serve as a model for how we'll do it for
the query reducer.