♻️(front) implement getResourceList as a generic, reusable saga
Instead of one saga by resource, make a single saga that will be
shared by all the resources; it will be the reducer's job to
discriminate incoming actions based on their resourceName property.
This allows us to make our code substantially simpler at the cost
of a little bit of boilerplate in the resource reducers - which we
can abstract away later on if it becomes an issue.