✨(front) wrap Search to fetch subjects & orgs on startup
Make a container for Search to connect it so it can dispatch
actions to GET all subjects & all orgs from the API on boot. This
is a bit crude as it does not limit the request to the necessary
data; we can get away with it as subjects and orgs are both pretty
light. We do it this way for simplicity's sake, and so we can display
course lists & filters as fast as possible.