🔧(front) enable typescript strict mode
TypeScript strict mode is not the same thing as "use strict" in
the browser context. Instead, it's an evolving flag for the
most comprehensive typechecking the compiler has to offer. Right now
it enables all existing options (including `noImplicitAny`, which we
had, and `strictNullChecks`, which had us fix some issues). Later on
any new options will be activated by default when we upgrade TS.