♻️(search) rename facets to use singular related record name
We originally named our first facets after the name of the property
on the course object, which is a plural as it contains an array.

However, such coupling is not necessary: the query for the ES
aggregation must match record property names but the aggregation's
name is arbitrary.

Using a singular makes using it much simpler on the client
side. Also our other aggregations will not necessarily involve
related record terms and would therefore have arbitrary names
too.