♻️(core) refactor "create_cms_data" command to be modular
This function was a bit of a mess and could not be tested. A
catch all test was implemented that took 30 seconds to execute.
I factorized a helper "create_i18n_page" that should be
usefull in other situations as well. The new tests make use of
mocks so that we can really test the details of our functions
while keeping the test suite fast.