♻️(front) clean up resource list params in reducer
As we're not caching list responses anymore, we don't need the query
key. On the other hand, we do need the current params to modify them
as users interact with the filters.
Set defaults and clean up APIListCommonRequestParams. It makes more
sense to have them specified and then use Partial<> when we accept
incomplete subsets.