🐛(api) fix an issue on org & subject filters on course search
- Expected Behavior
A call on `/api/v1.0/course?organizations=345` should filter courses
to keep those that belong to organization 345.
-> `CourseListForm` should provide the value as `[ '345' ]`

- Actual Behavior
The call filters courses that belong to organizations 3, 4 and 5.
-> `CourseListForm` provides the value as `'345'`

- Explanation and solution
This was due to the way we sidestepped the magic list-eating behavior of
QueryValueDict: we patched the values on the form after initializing it
with an object that contains values taken from QueryValueDict
with `v[0] if len(v) == 1 else v`.

Our own code replaces lists of 1 element with the element when the list
is one element long. That is the expected behavior, *except* for
`organizations` and `subjects`.
Patching them after the the form is initialized is too late: the form data
is already set and we're not updating it.

Simply patch the object we then pass to `CourseListForm`. This way the form
receives (and outputs) the expected values.