👷(linting) add black the uncompromising code formatter
We are losing a lot of time during reviews merely on
code formatting neatpicking. Black can check and even
fix formatting to always follow the exact same rules.
Black formatting rules are designed for minimal diffs
which eases code reviews.

We had to loosen up pylint's rules on indentation
alignment and similarities because of the way Black
is formatting the code.