♻️(front) pluralize resource names
We need a standard name for our resources (course VS courses, etc.)
so we do not have to convert between the singular and plural forms
depending on the place and mode we need to use their names.
This commit handles that renaming and standardization throughout
the frontend code: state keys, objects & type names, and of course
any interaction point with the backend.