🐛(search) fix subjects+organizations terms queries in the same query
The "terms" query does not support multiple different keys on the
same query. This is not documented but the issue arose when we attempted
it on our live ES instance.
Use a "bool" query that allows us to combine multiple queries (and therefore
2 or more "terms" queries) into one. It's also structurally cleaner as
it allows us to express our constraints as arrays of invididual constraints
with a filter/ranking/necessity marker.