✨(front) add filters to search when the user clicks on a filter element
This is the first (and most essential) functionality we need to implement:
add a filter to the current search when a user clicks on the corresponding
filter element in the list on the left-hand side of the screen.
Rely on SearchFilterGroupContainer to do the heavy lifting while
SearchFilterGroup & SearchFilter only pass along values and events.