👌(front) simplify filter updates computation
The first version was too complicated, even though it satisfied the
compiler completely. We needed to find a clearer way to write it, but
the simplest for humans (reverse switch + hash table) was not unterstood
by the compiler.
This is a compromise that is simpler to read for humans and still fully
typechecks.