🐛(front) use undefined instead of null to remove QS values
The lib we use to stringify query strings handles null & undefined
differently: null keeps the key but does not assign any value to it,
where as undefined removes the key.
Our API does not expect valueless keys to be present. Use undefined
instead of null so our requests comply with our API.
Additionally, this also makes our URLs cleaner as it looks better to
not leave artifacts of removed filters behind.