🔧(toolbar) allow to configure position of toolbar menu items
There is no easy way to determine automatically the position at
which we should insert the page extension menu items so that
they are positioned below the item to edit a page. We should
make it a setting so it can be ajusted depending on what other
menu items are present on an instance.