✨(autosuggest) implement a SearchSuggestion type for assisted search
We need a "container" type as search suggestions can be of several
types: courses, organizations, subjects, later on maybe people or
other things.
This type allows us to wrap the values themselves and add type
information so we can easily and cleanly use them. The suggestHumanName
util provides one such clean way to extract the human readable
text representation of a search suggestion.