🐛(search) do not apply dimension D's filter when faceting on it
We were having an issue where counts on facets were not what we expected,
as counts for eg. the organizations facets were only including courses
that matched the current organizations filter.
This made the facet counts pretty much useless for us. We can instead
facet on everything and use a specific filter for each aggregation,
that reproduces the main query but omits the current dimension.