🐛(front) fix an issue that prevented the 1st filter from being added
An overlooked parameter in a partial application in
SearchFilterGroupContainer led to the wrong value being passed into
our updateFilter helper from the filter. As a result, it was not
possible to add a value to the filter when its initial value was
undefined.
Adding the currentParams to the partially applied parameters on
updateFilters in SearchFilterGroupContainer ensures it has all the
data it needs to do update the filters.