🐛(front) fix an issue where a race condition caused filters to break
In some cases, depending on the order of arrival of data between
courses facets and organizations & subjects values, the filters
were breaking as SearchFilterGroupContainer through getFilterFromState
was trying to access properties on records that did not exist yet.
Using a safe check and filtering out empty values is enough to
prevent this from breaking the filters.

NB: the Maybe monad in the state shape should help us ensure we
don't run into this error again.