✅(front) add integration tests for the filters
The basic filter functionality was broken in a recent refactor:
two separate bugs could cause the filters to not be shown at all, or
make it impossible to add or remove filters.

This is not acceptable. Also, those bugs could not be found through
unit tests alone.
Add a filters integration test (the first of many?) to secure this
part of the code.