♻️(filters) improve filter definition typings
Resource based filter definitions were loosely typed in the reducer
and suddenly typed as FilterDefinitions when we needed them.

Rework that as a bare FilterDefinition type and an extending type
with the value property. This way we get to type the resource based
ones as FilterDefinition in the state and cleanly move to the
FilterDefinitionWithValues after going through getFilterFromState.

We also get to specify if other functions and components expect a
simple FilterDefinition or if they need the values too.