✚(autosuggest) add a dependency for search autosuggestion in React
The react-autosuggestion is full-featured, fully customizable (it comes
without any styling) and extensible in all the ways we might want.
It basically provides the autosuggestion UI as decoupled from any
other concerns as possible.
Use some bootstrap styles so we get a usable component without
writing our own CSS on top for now.