✨(courses) use Search as a template for courses in the demo-site
Search is a template that uses our front code to display a list of
courses along with filters and a search field. It is in fact the
intended template to show the user a list of courses.
Update create_demo_site so it sets up that template for the
courses page.