♻️(front) remove default exports from all our TS modules
In our opinion default exports inside app code are a bad practice
as it decouples the names of imported functions from the names of
the exports, which is not real decoupling as the underlying functionality
is still very much linked.
Using default exports has no benefit we can see but makes it easy to
transparently import the wrong module.
Just remove them and use only named exports/imports.