✨(person) add plugin to display a person on CMS pages
The person plugin will represent a person page extension
on other pages. The content rendered by the plugin is
built with the content inserted in the placeholders of
the original person page. The idea is that other teams
can customize the content of each page and plugin without
having to modify models and database schemas.