♻️(tests)refactor filter integration test
We used both `mount` (for full rendering) and `render` (for static
rendering to actual update our components), which was awkward as they
return wrappers with different possibilities and APIs.
Just use `mount` everywhere and wrap it in a function call so we
get freshly updated wrappers everytime we need to evaluate an action
or perform something.