🐛(circle) fix yarn.lock file update
Only install Greenkeeper-lockfile if it is not already installed.
Previous implementation using "yarn add" was updating this dependency
(and the lockfile), this was not intended as the resulting lockfile did
not match the PR new package.json file.

Once the yarn.lock file has been updated by greenkeeper, we need to save
this new version in the job workspace to restore it during the
installation. Using the '--frozen-lockfile' option of 'yarn install'
ensures that our new lockfile is properly correlated with dependencies
definition in the package.json file.