🐛(circle) force lockfile-front as the first job
As mentionned in greenkeeper-lockfile documentation [1], the
lockfile-front job SHOULD be the FIRST job to run. I misunderstood this
rule and thought that the first job of the front-end workflow would be
sufficient. This was a mistake: as we have multiple parallel workflows,
all other workflows' first job SHOULD require the lockfile-front job to
ensure that it will always be the first job to run.

[1] https://github.com/greenkeeperio/greenkeeper-lockfile#circleci-workflows