🔧(typescript) update typescript and its config to import JSON
We need to update typescript to take advantage of its new capability
to import and get types from JSON files.

We previously solved the JSON import issue with an ambient type
declaration but that was limited as we can't import our app's
types into that ambient module. We therefore would have had to
either duplicate our types there or to make them ambient, which
would be worse than cleanly importing them.

Due to esModuleInterop, we can (must) also use default imports
with modules that expect babel-style synthetic default imports,
instead of resorting to the TS-only hack of doing "* as whatever".

Enable the new "resolveJsonModule" option we upgraded TS to 2.9
for. It has a dependency on "allowSyntheticDefaultImports" which
is optional in TS but a standard behavior in babel and allows us
to import node-style broken exports as though they were default
imports.