🐛(search) show the proper organization for courses
Courses can have several linked organizations, listed in the API
on the `organizations` field. We operated on the assumption that
the "main" organization, the one that should be displayed in search
results, was the first in the list.
That assumption was wrong in our current use, and also encodes some
behavior implicitly. Use a specific field `organization_main` to
convey that information instead.