♻️(search) define frontend filters in settings.json
In order to make it easier to extend/modify the filters
available on the search courses view, define them in a JSON
settings file that can be extended and/or replaced, instead
of hardcoding them in out TypeScript code directly.

We need the settings.ts file to have more precise types than
what the compiler could get by itself from the source JSON
file. It also serves as a simple description of what's expected
of this JSON file.