🎨(seo) allow customizing header level in plugins
When a plugin includes content that is important for
SEO, we want to be able to add some hx headers to their
markup. The level that this <hx> element should get
depends on the page structure outside of the plugin ie
where the placeholder receiving the plugin is placed in
the page. We achieve this by setting a template variable
"header_level" before rendering the placeholders so that
we can derive the level of headers in the plugin from the
level of headers in their parent page.