✅(subject) enhance subject factory
Subject factory lacked to implement banner, logo and description, also
it was not included in demonstration site creation.

This has been performed with banner, logo and description field
implementation to the Factory, adding test to cover it and finally included
Subject batch creation inside 'create_demo_site' commandline. Used image files
for banner and logo are the same from Organization.