🔧(babel) add babeljs transpiler to our build
Babel allows us to automatically transpile our
code for the browsers we target and use autopolyfilling
so we can use proper JS and be sure it will run where
we need it to, while keeping our footprint at small as
possible.
It will also enable us to take advantage of automatic string
extraction when we add react-intl.

We needed to make small adjustments to the tsconfig so it
outputs ES6 code with ES6 modules that babel can then
pickup.