📦(richie) add base richie packaging suite
Following python packaging best practices, we have handcrafted a
setup.cfg file to package richie as a Django application. This was
required in order to be able to install it during our container image
builds and development.

TODO: dependencies are pinned and exhaustive (even secondary
dependencies). This will be improved in a follow-up pull request.