♻️(settings) move from JSON settings to TS settings
Settings are generally easier to wield when they are written
in the same language as the code that uses them. This is proving
true here as we need to automatically extract MessageDescriptors
from our settings. This will be a lot easier to do with TS settings
than with JSON.
Also, it makes consuming those settings a breeze throughout the app.