🔧(i18n) make an i18n folder to tidy up strings and translations
We'll have some i18n related stuff that needs to live in the repo:
we already have extracted strings from the frontend, the related POT
file, and exports of translations in PO format.

Create an i18n folder to aggregate them and gitignore extracted
strings as they can be regenerated from the code at any moment.