🔧(i18n) generate a POT from strings extracted from frontend code
The react-intl babel plugin we use for automatic string extraction
generates JSON files with simple Message Descriptors.
This is not a format that can be processed by the Crowdin (our
translation mangement SaaS tool).

Add a simple dependency to generate a standard POT file from
those extracted Message Descriptors. It will be called in the
CI or can be called directly through yarn (see its docs).