🚚(persons) move plugin test file to new tests directory
The project was refactored to make Richie a reusable app and the
tests were taken out of the app and to the root of the project.
The PR to add a person plugin was prepared before this change
and was mistakenly adding a test file inside the persons app
inside Richie.