🐛(persons) fix the person plugin to propose only person pages
When adding a person plugin to a page, the select box to choose
a person page was proposing all the pages in the CMS, not just
the person pages. Adding a "limit_choices_to" attribute on the
PersonPluginModel solves the issue as evidenced by the test
included with the fix.

Fixes #278