🔧(i18n) upload strings to crowdin through a node script
We opted not to use the Crowdin CLI due to it being a Java
dependency and requiring us to add a Java image to our container
setup, adding a lot of complexity to the project.
Instead, we can maintain a handful of small scripts to handle
what communications we need with the Crowdin API directly.