✨(i18n) generate a POT from backend strings
We extract i18n messages to a "locale" directory at the root
of the richie app containing the strings for all its subapps.
Keep django defaults so consumers have an easier time handling
i18n and do not run into trouble due to specific configuration
choices on our side.

We generate only the .pot file and added it to .gitignore.
The .po files will be pulled from Crowdin after completion
and committed in the "locale" directory.