🐛(models) fix getting linked page extensions
Many to many relations between page extenions are copied
when a page is published. We end up with links between
draft and published versions of the page extension but
the draft version should serve as the reference. I added
tests to evidence the bug and then fixed it.

TODO: add documentation describing these draft/published
challenges and how it works in Richie.