♻️(search) stop instantiating indexers
Indexers do not need to be instantiated: we do not need the concept
of an "indexer instance". The main reason to have them be classes
is to benefit from the extensibility it enables.
If every method is either a @staticmethod or a @classmethod, we can
stop instantiating them entirely, which makes them simpler to reason
about.