♻️(search) build ES requests in an Indexer method
We want to enable downstream developers to modify/extend/replace
search behavior as much as possible without specifying complex
interfaces between their code and ours.
In this case, the simplest way to let them customize ElasticSearch
queries is to delegate request building to the indexer so it can
be overridden to change the query behavior.