♻️(search) move facets list from course viewset todefaults
We need to know the resource (term) based facet keys in
both the indexer and the viewset. Also, it should be
possible to remove them to avoid issues when one does not
used subjects and/or organizations.
Get them from defaults (and a setting) so they can be shared
and easily customized/disabled.