♻️(person) simplify person plugin
We were retrieving all plugins related to a page and injecting
them in the context to render a plugin. In fact there is a much
simpler way to do it with the {% show_placeholder ... %}
templatetag.

TODO: the only case in which the new pattern is less good is to
render an image because we can't resize it differently for each
use case. But there is a solution that should be implemented in
another PR: pass the desired width/height in the context and
modify the template of the picture plugin to take into account
the values eventually passed in the context when generating the
thumbnail.