♻️(person) change person plugin to point to person instead of page
In a previous PR and For performance reasons, I had insisted to point
the person plugin to the related page object instead of pointing to
the person page extension. Playing with the person plugin, I realized
that the experience for page editors would be better if we linked to
the person page extension because we can more directloy access the
person firstname/lastname/title. It may also be more efficient in
the end because the number of person page extensions is much less
than the number of pages.