✨(search) set up our default filters on the backend side
We need to specify default filters to enable development & demo
using the sandbox site, and also to make sure they can be
overridden easily through settings. Use the defaults pattern
to ensure both these needs are met.
We wrote a basic suite of defaults filters that should be enough
for anyone to bootstrap a website around openedx and can be
customized later on.