🐛(demo) remove dead code in test for create demo site
An earlier version of the create_demo_site command was calling
the "create_i18n_page" helper to create the pages related to
our custom page extensions (course, organization, subject, person).
It has since be removed and the pages are created directly by
the factories. The test was not failing but some code was never
run so we can remove it.