✨(search) add courses list sorting to CoursesIndexer & ViewSet
Inside a filtered list of courses, we want to order courses by
their status: closed courses are of little interest to most people,
as are courses you can no longer enroll in.
On the other hand, we want to put open courses in which you can
enroll at the top of results.

Put all the relevant code on the indexer so it's easy to override
for implementors.