🎨(tests) improve names of test methods
It is good if all the names of our test methods follow
the same pattern. A good practice is to describe the
test uniquely from the most generic element to the most
specific. It brings clarity and allows smartly targetting
a group of tests with the "-k" option in pytest.