🐛(courses) public courses have the course runs of their draft version
The bug was that when publishing a draft course with course runs, the
course runs were not accessible any more from the published course.
This is because course runs are basic Django models, not DjangoCMS
page extensions... so they don't have draft/public versions. We could
have copied the relation between a draft course and its course runs
to the public course but it would have required a ManyToMany field
and more complexity when in fact we don't want to duplicate the course
runs. So we chose to have course runs attached only to the draft version
of a course and abstract fetching the course runs so that for a public
course we go and fetch the course runs from the related draft course.