✅(demo) improve demo site page tree
Page tree is just a flat list of pages, all enabled in navigation.
But we don't want every pages to be in menus and need a more advanced
architecture to be able to create an additional menu for annexe page
like About, Help, Legal mentions, etc..

This commit hides some pages from menus and add a new page "Annexe"
where "About" page is moved in.