♻️(courses) improve the publication mechanism for many to many fields
The many-to-many relations between courses and their organizations/
subjects were always pointing to the draft instance. This was
causing an issue on the following edge case:

- create a subject and publish it,
- add a course and link the subject to it,
- publish the course.

Now the published course should be on the published subject page
but it's absent.

Solving this edge case with the current relationships was convoluted
because it required adding the missing link between the published
subject and the draft course upon publishing the course. We ended up
updating the relationships for both the draft and the published
course. This was not necessarily a performance issue because the
number of organizations and subjects for a course is expected to be
small but this was hard to understand and lead to :hankey: code.

We decided to refactor the way relationships are made.

A draft course is still linked to its draft organizations and draft
subjects. But when we publish the course, the published course is now
linked to the published organizations corresponding to the draft
organizations that the draft course had (same for subjects).

CAVEAT: There is a new edge case but it is much less annoying so we
propose to merge this fix first and solve it in a separate PR:
- create a course and publish it,
- now create a draft subject,
- Edit the course to add the subject to it but don't publish the
  course yet,
- publish the subject

The published course now points to the subject which is not
expected because "adding the subject to the course" was still
a draft attempt.