💄(layout) finalize course glimpses
Currently course glimpses from organization and subject detail page
are a simple list of course title. It should be more elaborate and
display as a pretty grid of cards. This is related to issue #228.

This turns the current course list as glimpses and making their
behaviors from Sass settings. To be consistent this will apply to
course list from organization detail, subject detail and course list.