👷(tools) add package.json/makefile commands to run prettier
Right now we don't have easy-to-use commands to run prettier in
write mode on the codebase (the CI just runs it in list-different mode).

Add a "yarn prettier" command that can be run directly by end-users
with local installs of yarn & dependencies, and a corresponding
lint-front-prettier in Makefile to enable users to run the same
from the container.

To keep a consistent Makefile commands suite, keep a "lint-front"
command that runs both prettier & tslint, like we have "lint-back"
to run every backend linter at once.