✨(courserun) add method attribute to retrieve state of dates
We need to have an easy way to know what is the status of a
course run ie whether it is open, closed, on-going, etc at the
current time. This will be used to display a call to action or
just some information in the list of course run on a course page.