🚸(demo) diversify course runs on the demo site
On the demo site, all course runs were created in the future.
We want to have course runs in different states: some open,
and some future, past, open or closed for enrollment, etc.
We chose the option to loosen the factory and had to fix a
few tests as a consequence.