🔧(front) configure jest & update babel config for jest
As we're moving to jest, we need to add a Jest configuration file
to manage our file locations & our compiling settings. Move from
.babelrc to babel.config.json so we can trivially require it in
our Jest config.

Note: we need the `transformIgnorePatterns` key to ensure lodash-es
goes through compilation. It is the only dependency we use that is
not precompiled by the library author.