🚚(front) rename testSetup so Jest does not pick it up
Jest runs each `*.spec.ts` file it finds in our richie-front dir
as a test suite. testSetup is not supposed to be run that way: it's
simply a utility that some of our tests need to load before they
run.

Renaming it to omit the `.spec` is enough so Jest does not attempt
to run it.