♻️(front) update some tests for Jest
Jest can use centralized mocks. We already use one for `react-intl`
in some of our other projects, and it is helpful here to make our
tests simpler.

Two other changes are a forgotten "key" prop on a list component,
and the removal of a "Response" call which does not exist as a
global object in Jest.