💄(front) improve look & accessibility for SearchFilter active state
We needed a large "close" icon so it would be visible enough to be
obvious event next to the facet count badges. Replace the `&times;`
with a simple SVG icon.

Also, it seems to us that the aria-pressed label make more sense as
all those filter buttons are effectively toggles for the filter name
that appears inside them. The aria-label "Close" can then be applied
to the icon itself to make its purpose apparent.