🐛(person) refactor person plugin to fix public/draft display
The person plugin was always displayed in its draft version, even
on public pages. We started by changing the foreign key from person
to the related page in order to rely less on page extensions and
as much as possible on pages. Then we modified the template and added
helper methods on the plugin to make sure the plugin is always displayed
in its correct version. The tests were completed to reflect our findings.