🚸(courses) simplify and standardize placeholder names
The placeholder names are showing in the right side bar where the
user manages plugins for a page. Their name did not follow the
same pattern and did not always match with the title of the
corresponding section on the page making it unnecessarily difficult
for the user to understand what each placeholder is doing.