✨(subject) add related courses on the subject detail page
This is a copy/paste of what is already done on the organization
detail page. It is easier than using Elasticsearch+React so it
is good to add it now and replace it later when we reach the
limitations of this query (performance? need sorting?).

Note: the minimum similar lines before a pylint error is raised was
increased because we feel it is not a problem to have repeated
portions of code in tests.