♻️(css) relocalize variable values next to the elements they're styling
We're putting styles in the file structure next to the components &
templates they are styling. This enables us to confidently delete/refactor
code and makes it easier to understand components as their whole code
is colocated.

All this is jeopardized if the variable values that make up our actual
app styles are centralized in the .scss folder. It also does not serve
any purpose as the `!default` flag means they can be overridden if
someone sets another value in their own variables file.