🔧(circle) update the prettier lint job to check style files
prettier can reformat and checks style files (CSS and SCSS). As
it is our tool of choice to have a common format on frontend
related files, there is no reason not to enable it for style files.

Simply change the CI job to also check style files. Rename it to
remove the "front" part as it also checks style files in the backend
folder.

Also: took this opportunity to rename the package.json script that
runs prettier --write as it was messing with the CI job.