💄(front) fix close icon display when filter text spans 2 lines
The close icon on filters was getting shrinked as a flex item when
the filter name was spanning two lines (and therefore larger as
well).

Using `flex-shrink: 0` lets us avoid this unwanted behavior. Also
took this chance to add a small margin-left for the badge & close icon
so text does not flow too close to them.