✨(front) limit resource-based filters to 10 displayed suggestions
When we're using facet counts to display a resource-based filter, we're
only showing the top 10 available values in the list. (We'll be adding
a UI element to filter using other values).

When there are no facet counts, we're displaying all possible values. This
is not ideal as it might make the UI jitter on load & looks bad / makes
it seem like there is a bug when we really are missing the facets.

Sample 10 random values from the full list in this case.